Refactor: Launch a job's processes in one go
Collect the process specs for a job (that is, a pipeline) and
launch them in one go instead of launching each one as we process
the parse tree.

This will make it easier to do job control in the terminal, where
we need to use a helper process to launch processes under a
different controlling tty.